+++ src/stories/containers/Finances/useFinances.tsx
@@ -1,8 +1,9 @@
+import { useMediaQuery } from '@mui/material';
 import { siteRoutes } from '@ses/config/routes';
-import { useBudgetContext } from '@ses/core/context/BudgetContext';
 import { useThemeContext } from '@ses/core/context/ThemeContext';
+import { useFlagsActive } from '@ses/core/hooks/useFlagsActive';
 import { useRouter } from 'next/router';
-import { useState, useEffect } from 'react';
+import { useState, useEffect, useMemo } from 'react';
 import { useSWRConfig } from 'swr';
 import useSWRImmutable from 'swr/immutable';
 import useBreakdownChart from './components/BreakdownChartSection/useBreakdownChart';
@@ -20,40 +21,85 @@ import {
   existingColorsDark,
   nameChanged,
 } from './utils/utils';
+import type { Theme } from '@mui/material';
+import type { NavigationBreadcrumb } from '@ses/components/Breadcrumbs/Breadcrumbs';
 import type { BreakdownBudgetAnalytic } from '@ses/core/models/interfaces/analytic';
 import type { Budget } from '@ses/core/models/interfaces/budget';
 
-export const useFinances = (budgets: Budget[], initialYear: string) => {
+export const useFinances = (budgets: Budget[], allBudgets: Budget[], initialYear: string) => {
+  const [isEnabled] = useFlagsActive();
+  const { isLight } = useThemeContext();
+  const isMobile = useMediaQuery((theme: Theme) => theme.breakpoints.down('tablet_768'));
   const router = useRouter();
-  const { allBudgets } = useBudgetContext();
+  // TODO: this should not be used, instead we should use different keys to use different data/params
   const { mutate } = useSWRConfig();
   const [year, setYear] = useState(initialYear);
-  const { isLight } = useThemeContext();
-  const levelPath = null;
+
+  const urlPath = Array.isArray(router.query.path) ? router.query.path.join('/') : router.query.path;
+  const codePath = urlPath ? `atlas/${urlPath}` : 'atlas';
+  const levelNumber = codePath.split('/').length;
+  const levelOfDetail = levelNumber + 1;
+
+  const currentBudget = allBudgets.find((budget) => budget.codePath === codePath);
+  const icon = currentBudget?.image;
+  const title = nameChanged(currentBudget?.name || '');
+
   const handleChangeYears = (value: string) => {
     setYear(value);
-    router.push(`${siteRoutes.newFinancesOverview}?year=${value}`, undefined, { shallow: true });
+    router.push(`${siteRoutes.finances()}?year=${value}`, undefined, { shallow: true });
   };
 
+  // TODO: we should be using only one query and refetch the data depending on the selected granularity
   const { data: budgetsAnalytics } = useSWRImmutable(
     'analytics/annual',
-    async () => getBudgetsAnalytics('annual', year, 'atlas', 2, budgets) as Promise<BreakdownBudgetAnalytic>
+    async () =>
+      getBudgetsAnalytics('annual', year, codePath, levelOfDetail, budgets) as Promise<BreakdownBudgetAnalytic>
   );
-
   const { data: budgetsAnalyticsQuarterly } = useSWRImmutable(
     'analytics/quarterly',
-    async () => getBudgetsAnalytics('quarterly', year, 'atlas', 2, budgets) as Promise<BreakdownBudgetAnalytic>
+    async () =>
+      getBudgetsAnalytics('quarterly', year, codePath, levelOfDetail, budgets) as Promise<BreakdownBudgetAnalytic>
   );
   const { data: budgetsAnalyticsMonthly } = useSWRImmutable(
     'analytics/monthly',
-    async () => getBudgetsAnalytics('monthly', year, 'atlas', 2, budgets) as Promise<BreakdownBudgetAnalytic>
+    async () =>
+      getBudgetsAnalytics('monthly', year, codePath, levelOfDetail, budgets) as Promise<BreakdownBudgetAnalytic>
   );
 
-  const routes = ['Finances'];
-  const trailingAddress = routes.map((adr) => ({
-    label: adr,
-    url: `${router.asPath}${year}`,
-  }));
+  // generate the breadcrumb routes
+  const { trailingAddressMobile, trailingAddressDesktop } = useMemo(() => {
+    const segmentedCodePath = codePath.split('/');
+    const trailingAddressDesktop: NavigationBreadcrumb[] = [];
+    // build the breadcrumb url
+    segmentedCodePath.forEach((item, index) => {
+      if (item === 'atlas') {
+        // it is the first level
+        trailingAddressDesktop.push({
+          label: 'Finances',
+          url: `${siteRoutes.finances()}?year=${year}`,
+        });
+      } else {
+        // it is a deeper level
+        trailingAddressDesktop.push({
+          label: nameChanged(
+            allBudgets.find((budget) => budget.codePath === segmentedCodePath.slice(0, index + 1).join('/'))?.name || ''
+          ),
+          url: `${siteRoutes.finances(segmentedCodePath.slice(1, index + 1).join('/'))}?year=${year}`,
+        });
+      }
+    });
+
+    const trailingAddressMobile = [...trailingAddressDesktop].reverse();
+    trailingAddressMobile[0] = {
+      ...trailingAddressMobile[0],
+      style: { color: isLight ? '#25273D' : '#D2D4EF' },
+    };
+
+    return {
+      trailingAddressDesktop,
+      trailingAddressMobile,
+    };
+  }, [allBudgets, codePath, isLight, year]);
 
   const allMetrics = getTotalAllMetricsBudget(budgetsAnalytics);
 
@@ -65,6 +111,7 @@ export const useFinances = (budgets: Budget[], initialYear: string) => {
 
   const colorsDark = generateColorPalette(180, budgets.length, existingColorsDark);
 
+  // All the logic required by the CardNavigation section
   const cardsNavigationInformation = budgets.map((item, index) => {
     const budgetMetric =
       budgetsAnalytics !== undefined && budgetsAnalytics[item.codePath] !== undefined
@@ -75,19 +122,22 @@ export const useFinances = (budgets: Budget[], initialYear: string) => {
       image: item.image || '/assets/img/default-icon-cards-budget.svg',
       title: nameChanged(item.name),
       description: item.description || 'Finances of the core governance constructs described in the Maker Atlas.',
-      href: `${siteRoutes.newFinancesOverview}/${item.codePath.replace('atlas/', '')}?year=${year}`,
+      href: `${siteRoutes.finances(item.codePath.replace('atlas/', ''))}?year=${year}`,
       valueDai: budgetMetric[0].budget.value,
       totalDai: allMetrics.budget,
+      code: item.code,
       color: isLight ? colorsLight[index] : colorsDark[index],
     };
   });
 
+  // if there too many cards we need to use a swiper on desktop but paginated on mobile
   const [loadMoreCards, setLoadMoreCards] = useState<boolean>(cardsNavigationInformation.length > 6);
-
   const handleLoadMoreCards = () => {
     setLoadMoreCards(!loadMoreCards);
   };
-  const cardsToShow = loadMoreCards ? cardsNavigationInformation.slice(0, 6) : cardsNavigationInformation;
+
+  // pagination only happens on mobile devices
+  const cardsToShow = loadMoreCards && isMobile ? cardsNavigationInformation.slice(0, 6) : cardsNavigationInformation;
 
   // All the logic required by the breakdown chart section
   const breakdownChartSectionData = useBreakdownChart(
@@ -112,7 +162,7 @@ export const useFinances = (budgets: Budget[], initialYear: string) => {
   // All the logic for the Reserve Chart
   // This should be calculate
 
-  const reserveChart = useReservesWaterFallChart(levelPath, budgets, allBudgets);
+  const reserveChart = useReservesWaterFallChart(codePath, budgets, allBudgets);
 
   // invalidate cache and refetch all sections when year changes
   useEffect(() => {
@@ -122,7 +172,13 @@ export const useFinances = (budgets: Budget[], initialYear: string) => {
   }, [mutate, year]);
 
   return {
+    isEnabled,
     year,
+    levelNumber,
+    icon,
+    title,
+    trailingAddressMobile,
+    trailingAddressDesktop,
     handleChangeYears,
     cardOverViewSectionData,
     router,
@@ -133,7 +189,6 @@ export const useFinances = (budgets: Budget[], initialYear: string) => {
     handleLoadMoreCards,
     makerDAOExpensesMetrics,
     expenseReportSection: expenseTrendFinances,
-    trailingAddress,
     reserveChart,
   };
 };
